// api/enum/process.enum.ts
export enum Process {
    getHotels = 101,
    getHotelsFilter = 117, 
    getHotelsForStart = 118,
    getHotelForId = 102,
    createHotel = 103,
    updatedHotel = 104,
    removeHotel = 105,
    
    getCalificacionForHotel = 106,
    getCalificacionForId = 107,
    createCalificacion = 108,
    updatedcalificacion = 109,
    removeCalificacion = 110,

    getUsuarios = 111,
    getUserForId = 112,
    createUsuario = 113,
    updatedUsuario = 114,
    removeUsuario = 115,

    login = 116,

    getimageForID = 119,
    createImageForID = 120,
    createImage = 121,
    updateImage = 122,
    removeImage = 123
}